requirements.txt
pytest==7.4.4
pytest-playwright==0.4.4
pytest-bdd==7.0.1
playwright==1.41.2
pyyaml==6.0.1
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.5.0
